requirements.txt
streamlit
google.generative 